solutions, switched from int to long

+++ Solution.java
@@ -3,5 +3,5 @@ public interface Solution {
 	// Returns a random solution 1 move away
 	public Solution rand_move();
 	// Returns the residue of a solution
-	public int residue();
+	public long residue();
 }
